Refactor creation of the DB with SQL script invoking.
The name "engine_database.py" doesn't give much context about what it does.

The function "create_db" in this module is very bad because: 1) it doesn't create db; 2) it returns engine, not db.
It is better to call this file utils.py for instance and refactor its content a bit.

I've refactored the process of DB creation in such a way that every single function does one and only one thing (see Single Responsibility Principle).

I've also refactored several parts in the code so it looks more Pythonic.